backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.5.3
psycopg2-binary==2.9.9
requests==2.31.0
stripe==7.6.0
gunicorn==21.2.0
twilio==8.10.0
python-dotenv==1.0.0
apscheduler==3.10.4
pywebpush==1.14.0
eventlet==0.33.3
